+++ guest/rust/examples/basics/skinmesh/src/client.rs
@@ -108,7 +108,7 @@ fn App(
         });
     }
 
-    FocusRoot::el([FlowColumn::el([
+    FlowColumn::el([
         FlowRow::el([
             Text::el("Capeira"),
             Slider {
@@ -153,5 +153,5 @@ fn App(
             })
             .el(),
         ]),
-    ])])
+    ])
 }